sw floattable: fix lost floating table right before a table from DOC
The bugdoc has a floating table, anchored in a paragraph that is hidden
via character formatting. The bugdoc also has a normal table. This leads
to 1 table in Writer, but 2 tables in Word.

We already have code that tries to make sure floating tables have a
suitable anchor, see the code in WW8TabDesc::CreateSwTable(), but that
checks for the case when the next node after a floating table would be
table (and not text), instead of the hidden character property.

Fix the problem by not creating the hidden char attribute in the first
place in SwWW8ImplReader::SetToggleAttr() in case the pool item would be
inserted at the paragraph start and we just inserted a floating table,
which makes the 2nd table visible in Writer as well.

This is for DOC, interesting when Word converts this document to DOCX,
then the hidden attribute is removed, so there this is not really needed.

(cherry picked from commit 79ddca4def81198e3eee42eca8aca42fef964c80)

Change-Id: I3a7411e6fcc318740bcbd4b0cde9f34134f384a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154090
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>