Revert "ww8import create a pagedesc if continuous section changes margins"
This reverts commit c5ed52b1cd6f22787c94bec035ceecf9e1da3271.

The reverted commit would insert a pagebreak on section breaks if the
margin before and after the section break was different. While this
means a table after the section break will get the correct margin, it
would also create a page which is very breaking - it shouldn't be the
case in documents that have multiple continuous section breaks in a
single page that a new page is created for each of those breaks if the
margins are different - there are documents where this happens and in
those documents a single page gets converted to multiple messy pages.
Thus the table margins being correct after a section break with a margin
change should be fixed in a different way.

Change-Id: Ia5d5b98ddbd021803a6a9f80762b97db5808509f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154565
Tested-by: Jenkins
Tested-by: Paris Oplopoios <parisoplop@gmail.com>
Reviewed-by: Paris Oplopoios <parisoplop@gmail.com>
(cherry picked from commit df00cd5cf77eb2322f53b80296862940ba815d9a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/154624
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>