tdf#120262 sw floattable: no split when none of first row fits the vert space
The final problem with the bugdoc is that half row of the second
floating table was still on page 1, while it should be fully on page 2.

The reason for this was that first we thought we can't move forward,
since GetIndPrev() returns nullptr for a table that's inside a fly
frame. Then we thought it's a good idea to split, even if the split
would move the entire first row to the next page.

Fix the problem by:

- In SwTabFrame::MakeAll(), handle split flys after calling
  GetIndPrev(), an indirect prev of the anchor has the same meaning in
  this context.

- In SwTabFrame::Split(), fail for split flys in case only half of the first
  row's first line would fit, which gives an opporunity to call
  MoveFwd().

- In SwTabFrame::MakeAll(), call MoveFwd() in a way that it's not a
  problem that the GetIndPrev() call in MoveFwd() gives us a nullptr.

- At this point we don't split the table, we move it forward, but an
  empty master remains on page 1. Fix that by improving
  SwFlowFrame::MoveSubTree() to mark empty flys for deletion, similar to
  how it does the same for sections.

- Finally avoid a layout warning in SwTabFrame::MakeAll(), it's OK to
  try to split in case our split fly has a follow, we can move there on
  split failure.

Also bring the test document closed to the original bugdoc, so we can
assert the content on both pages.

(cherry picked from commit 45574624ff05673d44f11cdbbbb49e1af599133e)

Change-Id: I2d3f88342d91b3e256bc41416a9afb274a9309d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155091
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>