sc: rename oox::xls::Color to oox::xls::XlsColor
It's annoying that the class name is so similar to the commonly
used class Color, even when it is in its own namespace. To avoid
unneccesary conflicts and confusion, it was renamed.

(Similar would be also good for oox::drawingml::Color)

Change-Id: I68dcaf951b7faf089cb2b13d94576007f40c441e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155109
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 7f265875ef784d717f5a534bdca2cd36fdce7ecb)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156102
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>