tdf#156349 sw floattable: fix caption insert for tables inside split fly frames
Inserting a long enough caption below the table inside a split fly
resulted in a layout loop.

An additional, second problem is that in case a split fly only has a
table, that can be mapped to a floating table in Word formats, but once
a caption is inserted after the table (but still inside the frame), then
this will be a normal shape, which can't be split.

Fix the problem by disabling the "is split allowed" property on the fly
in case it's on, and only then insert the caption.

This not only avoids the layout loop, it also results in rendering that
will match after save + load, for Word formats.

(cherry picked from commit 107de1a2c6882213cf0ef6783417302f43cdada0)

Change-Id: Ie8b04e6266dd5c4068e5add86efb09cc6a6ea01f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155334
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>