tdf#77760 sw floattable: add support for footnotes, doc model
The bugdoc had a floating table, but we didn't convert the inline table
to floating because it had a footnote, and having footnotes in fly
frames is not supported in Writer.

One benefit of not having footnotes in fly frames is that once these are
saved to Word shapes, those can't have footnotes, either -- there is
value in limiting the UI to the constructs that can be also saved in
Word formats.

Fix the problem by allowing this for split flys, but not for
flys/headers/footers in general.

This is just the doc model, later steps (layout, etc) still need doing
in follow-up commits.

(cherry picked from commit 56da1d30afe48cc4acd79567052a575e81f8c7a0)

Conflicts:
	sw/qa/core/txtnode/txtnode.cxx

Change-Id: Ic3ba95877d7fad3f5f30743b89e2cf90e0d0af52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155963
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>