sc: add ODF import/export of the Theme + tests
One missing thing is the support in calc to save the Theme into
the ODS document and read that back. The theme element is added as
a child element to the office:styles - the same as it already is
added in Writer.

Also adds "Theme" property as a top level document property
so it is possible to get and set the theme in xmloff.

Also tests have been added to cover this usecases.

Change-Id: Ic214ff5e945b77d50e6c881def9d49509560a0e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156363
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit b36f7769dd07a6b6f55cdf4ce76e2f39ed186e89)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156456
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>